+++ dbos/_conductor/conductor.py
@@ -1,46 +1,15 @@
-import json
 import threading
 import time
 import urllib.parse
-from dataclasses import asdict, dataclass
-from enum import Enum
 from typing import TYPE_CHECKING, Optional
 
 from websockets import ConnectionClosed, ConnectionClosedOK
 from websockets.sync.client import ClientConnection, connect
 
+from . import protocol as p
+
 if TYPE_CHECKING:
     from dbos import DBOS
-
-
-class MessageType(str, Enum):
-    EXECUTOR_INFO = "executor_info"
-
-
-@dataclass
-class BaseMessage:
-
-    @classmethod
-    def from_json(cls, json_str):
-        data = json.loads(json_str)
-        kwargs = {k: v for k, v in data.items() if k in cls.__annotations__}
-        return cls(**kwargs)
-
-    def to_json(self):
-        dict_data = asdict(self)
-        return json.dumps(dict_data)
-
-
-@dataclass
-class ExecutorInfoRequest(BaseMessage):
-    type: MessageType = MessageType.EXECUTOR_INFO
-
-
-@dataclass
-class ExecutorInfoResponse(BaseMessage):
-    executor_id: str
-    application_version: str
-    type: MessageType = MessageType.EXECUTOR_INFO
 
 
 class ConductorWebsocket(threading.Thread):
@@ -62,9 +31,9 @@ class ConductorWebsocket(threading.Thread):
                 self.websocket = connect(self.url)
                 while not self.evt.is_set():
                     message = self.websocket.recv()
-                    base_message = BaseMessage.from_json(message)
+                    base_message = p.BaseMessage.from_json(message)
                     type = base_message.type
-                    if type == MessageType.EXECUTOR_INFO.value:
+                    if type == p.MessageType.EXECUTOR_INFO.value:
                         pass
                     else:
                         self.dbos.logger.warning(f"Unexpected message type: {type}")
